+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- Fix bug when unsubscribing different stream from the same channel ([@amree][])
+
 ## 1.6.0-dev
 
 - Refactor WebSocket message writing to use queues instead of channels to better handle slow clients. ([@palkan][])